+++ package.json
@@ -38,7 +38,7 @@
     "@supabase/supabase-js": "^2.49.1",
     "@t3-oss/env-nextjs": "^0.12.0",
     "@tailwindcss/postcss": "^4.0.8",
-    "@tanstack/react-query": "^5.66.9",
+    "@tanstack/react-query": "^5.67.2",
     "@tanstack/react-query-devtools": "^5.66.9",
     "@tanstack/react-table": "^8.21.2",
     "@types/three": "^0.174.0",
